delete route user delete removes all the sub tasks recorded by him

+++ src/routers/task.js
@@ -1,11 +1,15 @@
 const express = require('express')
 const router = new express.Router()
-const Task=require('../models/task')
+const Task = require('../models/task')
+const auth=require('../middleware/auth')
 
 // Task endpoints
-router.post('/tasks', async (req, res) => {
-    const task = new Task(req.body)
+router.post('/tasks', auth,async (req, res) => {
 
+    const task = new Task({
+        ...req.body,
+        owner:req.user._id
+    })
     try {
         await task.save()
         res.status(201).send(task)
@@ -14,12 +18,14 @@ router.post('/tasks', async (req, res) => {
     }
 
 })
-router.get('/tasks', async (req, res) => {
+router.get('/tasks',auth, async (req, res) => {
     
 
     try {
-        const allTask = await Task.find()
-        res.status(201).send(allTask)
+        await req.user.populate('tasks')
+        // ?Another Way 
+        // const allTask = await Task.find({owner:req.user._id})
+        res.status(201).send(req.user.tasks)
     } catch (error) {
         res.status(502).send(error)
     }
@@ -27,12 +33,18 @@ router.get('/tasks', async (req, res) => {
 })
 
 // ? Using async/await syntax kills the promises chain
-router.get('/tasks/:id', async (req, res) => {
-
+router.get('/tasks/:id', auth,async (req, res) => {
+    const _id = req.params.id
+    // user can fetch the data that he created, else is not going to.
     try {
         
-        const task = await Task.findById(req.params.id)
-        res.status(201).send(task)
+        // const task = await Task.findById(req.params.id)
+        const task = await Task.findOne({ _id, owner: req.user._id })
+        if (!task) {
+            return res.status(404).send()
+
+        }
+        res.send(task)
     } catch (error) {
         res.status(402).send(error)
     }
@@ -51,21 +63,24 @@ const fillAbleTaskInput = (input) => {
     return isValid
 }
 
-router.patch('/tasks/:id', async (req, res) => {
-    const reqFilled = Object.keys(req.body)
+router.patch('/tasks/:id',auth, async (req, res) => {
+    const reqFill = Object.keys(req.body)
     const fillAble = ['description', 'completed']
-    const isValid = reqFilled.every((x) => fillAble.includes(x))
+    const isValid = reqFill.every((x) => fillAble.includes(x))
     // Check fillable function
     if (!isValid) {
         return res.status(400).send({ error: "invalid updates" })
     }
     try {
-        const task = await Task.findByIdAndUpdate(req.params.id, req.body)
-        reqFilled.forEach((value) => { reqFilled[value] = req.body[value] })
-        task.save()
+        // find by the owner value
+        const task=await Task.findOne({_id:req.params.id,owner:req.user._id})
+        // const task = await Task.findByIdAndUpdate(req.params.id, req.body)
+       console.log(task);
         if (!task) {
             return res.status(404).send()
         }
+        reqFill.forEach((value) => { reqFill[value] = req.body[value] })
+      await task.save()
         res.send(task)
     } catch (error) {
         res.status(400).send({ error })
@@ -77,11 +92,11 @@ router.patch('/tasks/:id', async (req, res) => {
 
 
 
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id',auth, async (req, res) => {
     try {
-        const taskOps = await Task.findByIdAndDelete(req.params.id)
+        const taskOps = await Task.findByIdAndDelete({ _id: req.params.id,owner:req.user._id})
         if (!taskOps) {
-            return res.status(404).send("We could not find the task with this id:", req.params.id)
+            return res.status(404).send(`We could not find the task with this id ${req.params.id}`)
         }
         res.send("Task deleted Successfully")
     } catch (error) {
